backend/requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
colorama==0.4.4
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
google-generativeai>=0.3.1
idna==2.10
iniconfig==1.1.1
packaging==21.0
Pillow>=10.0.0
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
urllib3==1.26.6
